- Added configure checking of libqwt
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@8082 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ OMPlot/OMPlotGUI/plotwindow.cpp
@@ -118,8 +118,8 @@ void PlotWindow::initializePlot()
     mpQwtPlot->replot();
 
     //plotpicker, Plotpanner, plotzoomer
-    mpPlotPicker = new QwtPlotPicker(QwtPlot::xBottom, QwtPlot::yLeft,
-                                     QwtPlotPicker::CrossRubberBand, QwtPicker::AlwaysOn,
+    mpPlotPicker = new QwtPlotPicker((int)QwtPlot::xBottom, (int)QwtPlot::yLeft,
+                                     (int)QwtPlotPicker::CrossRubberBand, (int) QwtPicker::AlwaysOn,
                                      mpQwtPlot->canvas());
     mpPlotPicker->setStateMachine(new QwtPickerDragPointMachine());
     mpPlotPicker->setTrackerPen(QColor(Qt::black));
